Stubbing out a "State"
This holds the current engine state, and so far will contain the raw
text. This object will likely evolve over time to provide a snapshot of
how things are (the input mode, selection, cursor location, etc.). We'll
also figure out the best way to convey this data to be mindful of memory
copies.